requirements.txt
weaviate-client>=4.4.0
python-dotenv>=1.0.0
cohere>=5.0.0
anthropic>=0.25.0
numpy>=1.24.0
tiktoken>=0.5.0
